requirements.txt
asgiref==3.3.1
astroid==2.4.1
cffi==1.14.3
cryptography==3.2.1
Django==3.1.3
psycopg2-binary>=2.8
django-debug-toolbar==3.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.5.0
dodgy==0.2.1
flake8==3.8.4
flake8-polyfill==1.0.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pep8-naming==0.10.0
prospector==1.3.1
protobuf==3.13.0
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==5.1.1
pyflakes==2.2.0
PyJWT==1.7.1
pylint==2.5.3
pylint-celery==0.3
pylint-django==2.1.0
pylint-flask==0.6
pylint-plugin-utils==0.6
pytz==2020.4
PyYAML==5.3.1
requirements-detector==0.7
sentry-sdk==0.19.5
setoptconf==0.2.0
six==1.15.0
snowballstemmer==2.0.0
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
pillow==7.2.0
celery==5.0.5
redis==3.5.3
django-redis==4.12.1
drf-yasg==1.20.0
raven==6.10.0
django-storages==1.11.1
boto3==1.16.57
fcm-django==1.0.4
django-cors-headers